one loop of simlod correct

+++ src/ooc_structures/structureUpdate.cpp
@@ -314,7 +314,8 @@ void createCudaMemory(CuRast* editor, CUcontext* context,
 		new_node->aabb_index = cur_node->aabb_index;
 
 		new_node->points_counter = cur_node->counter.load();
-		new_node->children_ids = cur_node->children_ids;
+		new_node->children_ids = uint32_t(cur_node->children_ids);
+		// new_node->children_ids = cur_node->children_ids;
 		new_node->children_visibility = cur_node->children_visibility;
 		new_node->level = level;
